Work on Turbo-Frame section.
Inline Editing with Turbo Frames.

---
We will edit our view files so that when we edit/delete a task inside the list,
we will receive a response in the form of a fragment of the corresponding page,
which will be framed in a turbo-frame.
---

1. Let's edit one of the existing Tasks. After we submit the form, if the form
validation succeed, we can see the updated Task instance immediately.
2. If we delete the Task, then in response we will receive a message of successful
deletion, which will take the place of the Task in the turbo-frame element.

---

Note that we have implemented the functions of creating, editing and deleting
Tasks without reloading the page. To do this, we did not need to write a single
line of code in our javascript file.