Work on Stimulus-Basics section.
---
Stimulus is designed to enhance static or server-rendered HTML by connecting
JavaScript objects to elements on the page using simple annotations.
These JavaScript objects are called controllers, and Stimulus continuously
monitors the page waiting for HTML data-controller attributes to appear.
For each attribute, Stimulus looks at the attribute’s value to find a
corresponding controller class, creates a new instance of that class,
and connects it to the element.
---

1. Add dependencies to package.json.
2. Create the first Stimulus controller.
3. Implement autoload the controllers with stimulus-webpack-helpers package.